+++ geotrek/tourism/filters.py
@@ -1,14 +1,18 @@
-from django.utils.translation import gettext_lazy as _
-
-from django_filters.filters import ModelMultipleChoiceFilter, ChoiceFilter
 import django_filters.rest_framework
-from django.db.models import Q
-from geotrek.authent.filters import StructureRelatedFilterSet
 from django import forms
 from django.utils.datetime_safe import datetime
+from django.utils.translation import gettext_lazy as _
+from django_filters.filters import ChoiceFilter, ModelMultipleChoiceFilter
 
-from .models import TouristicContent, TouristicEvent, TouristicContentType1, TouristicContentType2
+from geotrek.authent.filters import StructureRelatedFilterSet
 from geotrek.zoning.filters import ZoningFilterSet
+
+from .models import (
+    TouristicContent,
+    TouristicContentType1,
+    TouristicContentType2,
+    TouristicEvent,
+)
 
 
 class TypeField(forms.ModelMultipleChoiceField):
@@ -79,18 +83,3 @@ class TouristicEventFilterSet(ZoningFilterSet, StructureRelatedFilterSet):
             'before', 'approved', 'source', 'portal', 'provider',
             'bookable', 'cancelled', 'place'
         ]
-
-
-class TouristicEventApiFilterSet(django_filters.rest_framework.FilterSet):
-    ends_after = django_filters.DateFilter(method='events_end_after')
-
-    class Meta:
-        model = TouristicEvent
-        fields = ('ends_after', )
-
-    def events_end_after(self, queryset, name, value):
-        if not value:
-            return queryset
-        return queryset.filter(
-            Q(end_date__isnull=True) | Q(end_date__gte=value)
-        )
