+++ geotrek/tourism/views.py
@@ -1,34 +1,26 @@
 import logging
-import os
 
 from django.conf import settings
 from django.contrib.auth.decorators import login_required
 from django.contrib.gis.db.models.functions import Transform
 from django.core.exceptions import PermissionDenied
-from django.db.models import Q, Sum
+from django.db.models import Sum
 from django.http import Http404, HttpResponse
 from django.shortcuts import get_object_or_404
 from django.utils.decorators import method_decorator
-from django.utils.translation import gettext as _
 from django.utils.html import escape
 from django.views.generic import CreateView
-from django_filters.rest_framework import DjangoFilterBackend
 from mapentity.views import (MapEntityCreate, MapEntityUpdate, MapEntityList, MapEntityDetail,
                              MapEntityDelete, MapEntityFormat, MapEntityDocument)
 from rest_framework import permissions as rest_permissions, viewsets
-from rest_framework.permissions import IsAuthenticatedOrReadOnly
-from rest_framework.renderers import JSONRenderer
-from rest_framework.response import Response
-from rest_framework.views import APIView
 
 from geotrek.authent.decorators import same_structure_required
-from geotrek.common.mixins.api import APIViewSet
 from geotrek.common.mixins.views import CompletenessMixin, CustomColumnsMixin
 from geotrek.common.models import RecordSource, TargetPortal
 from geotrek.common.views import DocumentPublic, DocumentBookletPublic, MarkupPublic
 from geotrek.common.viewsets import GeotrekMapentityViewSet
 from geotrek.trekking.models import Trek
-from .filters import TouristicContentFilterSet, TouristicEventFilterSet, TouristicEventApiFilterSet
+from .filters import TouristicContentFilterSet, TouristicEventFilterSet
 from .forms import TouristicContentForm, TouristicEventForm, TouristicEventOrganizerFormPopup
 from .models import (TouristicContent, TouristicEvent, TouristicContentCategory, TouristicEventOrganizer, InformationDesk)
 from .serializers import (TouristicContentSerializer, TouristicEventSerializer,
@@ -167,25 +159,6 @@ class TouristicContentViewSet(GeotrekMapentityViewSet):
         return qs
 
 
-class TouristicContentAPIViewSet(APIViewSet):
-    model = TouristicContent
-    serializer_class = TouristicContentAPISerializer
-    geojson_serializer_class = TouristicContentAPIGeojsonSerializer
-
-    def get_queryset(self):
-        qs = TouristicContent.objects.existing()
-        qs = qs.filter(published=True)
-
-        if 'source' in self.request.GET:
-            qs = qs.filter(source__name__in=self.request.GET['source'].split(','))
-
-        if 'portal' in self.request.GET:
-            qs = qs.filter(Q(portal__name=self.request.GET['portal']) | Q(portal=None))
-
-        qs = qs.annotate(api_geom=Transform("geom", settings.API_SRID))
-        return qs
-
-
 class TouristicEventList(CustomColumnsMixin, MapEntityList):
     queryset = TouristicEvent.objects.existing()
     filterform = TouristicEventFilterSet
@@ -320,41 +293,6 @@ class TouristicEventViewSet(GeotrekMapentityViewSet):
         return qs
 
 
-class TouristicEventAPIViewSet(APIViewSet):
-    model = TouristicEvent
-    serializer_class = TouristicEventAPISerializer
-    geojson_serializer_class = TouristicEventAPIGeojsonSerializer
-    filter_backends = [DjangoFilterBackend, ]
-    filterset_class = TouristicEventApiFilterSet
-
-    def get_queryset(self):
-        qs = TouristicEvent.objects.existing()
-        qs = qs.filter(published=True)
-
-        if 'source' in self.request.GET:
-            qs = qs.filter(source__name__in=self.request.GET['source'].split(','))
-
-        if 'portal' in self.request.GET:
-            qs = qs.filter(Q(portal__name=self.request.GET['portal']) | Q(portal=None))
-
-        qs = qs.annotate(api_geom=Transform("geom", settings.API_SRID))
-        return qs
-
-
-class InformationDeskViewSet(viewsets.ModelViewSet):
-    model = InformationDesk
-    queryset = InformationDesk.objects.all()
-    serializer_class = InformationDeskGeojsonSerializer
-    permission_classes = [rest_permissions.DjangoModelPermissionsOrAnonReadOnly]
-
-    def get_queryset(self):
-        qs = super().get_queryset()
-        if self.kwargs.get('type'):
-            qs = qs.filter(type_id=self.kwargs['type'])
-        qs = qs.annotate(api_geom=Transform("geom", settings.API_SRID))
-        return qs
-
-
 class TrekInformationDeskViewSet(viewsets.ModelViewSet):
     model = InformationDesk
     serializer_class = InformationDeskGeojsonSerializer
@@ -456,37 +394,3 @@ if 'geotrek.diving' in settings.INSTALLED_APPS:
             if 'portal' in self.request.GET:
                 queryset = queryset.filter(portal__name=self.request.GET['portal'])
             return queryset.annotate(api_geom=Transform("geom", settings.API_SRID))
-
-
-class TouristicCategoryView(APIView):
-    """ touristiccategories.json generation for API """
-    renderer_classes = (JSONRenderer,)
-    permission_classes = (IsAuthenticatedOrReadOnly,)
-
-    def get(self, request, format=None, lang=None):
-        response = []
-        content_categories = TouristicContentCategory.objects.all()
-
-        if request.GET.get('categories', False):
-            categories = request.GET['categories'].split(',')
-            content_categories.filter(pk__in=categories)
-
-        for cont_cat in content_categories:
-            response.append({'id': cont_cat.prefixed_id,
-                             'label': cont_cat.label,
-                             'type1_label': cont_cat.type1_label,
-                             'type2_label': cont_cat.type2_label,
-                             'pictogram': os.path.join(settings.MEDIA_URL, cont_cat.pictogram.url),
-                             'order': cont_cat.order,
-                             'slug': 'touristic-content'})
-
-        if request.GET.get('events', False):
-            response.append({'id': 'E',
-                             'label': _("Touristic events"),
-                             'type1_label': "",
-                             'type2_label': "",
-                             'pictogram': os.path.join(settings.STATIC_URL, 'tourism', 'touristicevent.svg'),
-                             'order': None,
-                             'slug': 'touristic-event'})
-
-        return Response(response)
